requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
djangorestframework-simplejwt>=5.0,<6.0
psycopg2>=2.9,<3.0
django-redis>=5.0,<6.0
redis>=4.0,<5.0
